refactor(trusty-mpm): move the `tm ls` connector out of session_picker
Rebasing #4965's `n [name]` picker command onto current main put
session_picker.rs at 524 SLOC. #5035 (`tm f`) added ~94 lines to the same
file after #4978 branched, consuming the headroom #4978's own legend
extraction had freed — main now sits at exactly 500, so any addition fails
check_line_cap.sh.

`should_show_picker` and `run_ls_connector` move verbatim to a new
`session_ls_connector` module. They are the `tm ls` orchestrator — deciding
whether to open the picker at all — not the picker itself, and they have one
call site. No logic changed; the fetch, filter, sort and picker-loop functions
they call still live in session_picker and are called from there.

session_picker.rs: 524 -> 460 SLOC.

Refs #4965
