+++ crates/trusty-mpm/src/bin/tm/commands/mod.rs
@@ -83,6 +83,7 @@ pub(crate) mod serve_stdio;
 pub(crate) mod services;
 pub(crate) mod sessctl;
 pub(crate) mod session;
+pub(crate) mod session_ls_connector;
 pub(crate) mod session_picker;
 pub(crate) mod session_picker_filter;
 pub(crate) mod session_picker_prune;
